requirements.txt
numpy
pygame
box2d-py